// src/app/modules/administracion/helpers/enums/kyc-actividad.enum.ts
export enum KYCEActividad {
    FIN = 0,
    NUEVO_FOLIO = 1,
    SOLICITUD = 2,
    CARGA_DOCUMENTAL = 3,
    VALIDACION_DIGITAL = 4,
    FIRMA_CLIENTE = 5,
    FIRMA_EJECUTIVO = 6,
    VALIDACION_FIRMAS = 7,
    CONFIRMACION_ENTREGA = 8,
    CONTACTO_TELEFONICO = 9,
    CONTACTO_ASEGURADORA = 10
}